fix: Resolve conftest import error in test_rhiza_workflows
Move helper functions (run_make, setup_rhiza_git_repo, strip_ansi) to a
dedicated helpers.py module to avoid pytest conftest discovery issues.
Python was finding tests/test_tiny_cta/conftest.py instead of the
intended tests/test_rhiza/conftest.py when using bare imports.
