+++ tests/test_rhiza/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for rhiza framework functionality."""
