+++ tests/test_rhiza/conftest.py
@@ -9,61 +9,15 @@ Provides test fixtures for testing git-based workflows and version management.
 import logging
 import os
 import pathlib
-import re
 import shutil
 import subprocess
 
 import pytest
 
-# Get absolute paths for executables to avoid S607 warnings
-GIT = shutil.which("git") or "/usr/bin/git"
-MAKE = shutil.which("make") or "/usr/bin/make"
+from tests.test_rhiza.helpers import GIT, run_make, setup_rhiza_git_repo, strip_ansi
 
-
-def strip_ansi(text: str) -> str:
-    """Strip ANSI escape sequences from text."""
-    ansi_escape = re.compile(r"\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])")
-    return ansi_escape.sub("", text)
-
-
-def run_make(
-    logger, args: list[str] | None = None, check: bool = True, dry_run: bool = True
-) -> subprocess.CompletedProcess:
-    """Run `make` with optional arguments and return the completed process.
-
-    Args:
-        logger: Logger used to emit diagnostic messages during the run
-        args: Additional arguments for make
-        check: If True, raise on non-zero return code
-        dry_run: If True, use -n to avoid executing commands
-    """
-    cmd = [MAKE]
-    if args:
-        cmd.extend(args)
-    # Use -s to reduce noise, -n to avoid executing commands
-    flags = "-sn" if dry_run else "-s"
-    cmd.insert(1, flags)
-    logger.info("Running command: %s", " ".join(cmd))
-    result = subprocess.run(cmd, capture_output=True, text=True)
-    logger.debug("make exited with code %d", result.returncode)
-    if result.stdout:
-        logger.debug("make stdout (truncated to 500 chars):\n%s", result.stdout[:500])
-    if result.stderr:
-        logger.debug("make stderr (truncated to 500 chars):\n%s", result.stderr[:500])
-    if check and result.returncode != 0:
-        msg = f"make failed with code {result.returncode}:\nSTDOUT:\n{result.stdout}\nSTDERR:\n{result.stderr}"
-        raise AssertionError(msg)
-    return result
-
-
-def setup_rhiza_git_repo():
-    """Initialize a git repository and set remote to rhiza."""
-    subprocess.run([GIT, "init"], check=True, capture_output=True)
-    subprocess.run(
-        [GIT, "remote", "add", "origin", "https://github.com/jebel-quant/rhiza"],
-        check=True,
-        capture_output=True,
-    )
+# Re-export for backwards compatibility
+__all__ = ["run_make", "setup_rhiza_git_repo", "strip_ansi"]
 
 
 MOCK_MAKE_SCRIPT = """#!/usr/bin/env python3
