+++ tests/test_rhiza/test_rhiza_workflows.py
@@ -17,7 +17,8 @@ import subprocess
 from pathlib import Path
 
 import pytest
-from conftest import run_make, setup_rhiza_git_repo, strip_ansi
+
+from tests.test_rhiza.helpers import run_make, setup_rhiza_git_repo, strip_ansi
 
 
 @pytest.fixture(autouse=True)
